Version 0.0.5

+++ pprpg/src/main/java/ppodds/rpg/pprpg/command/lvl.java
@@ -10,11 +10,11 @@ import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
-import org.bukkit.craftbukkit.v1_10_R1.entity.CraftPlayer;
+import org.bukkit.craftbukkit.v1_11_R1.entity.CraftPlayer;
 import org.bukkit.entity.Player;
 
-import net.minecraft.server.v1_10_R1.IChatBaseComponent.ChatSerializer;
-import net.minecraft.server.v1_10_R1.*;
+import net.minecraft.server.v1_11_R1.IChatBaseComponent.ChatSerializer;
+import net.minecraft.server.v1_11_R1.*;
 import ppodds.rpg.pprpg.PPRPG;
 import ppodds.rpg.pprpg.mysql.MySQL;
 import ppodds.rpg.pprpg.skill.mana.Mana;
@@ -62,7 +62,7 @@ public class lvl implements CommandExecutor
 				
 				sender.sendMessage(ChatColor.translateAlternateColorCodes('&', "&2==============================&f基本屬性&2=============================="));
 				sender.sendMessage(ChatColor.translateAlternateColorCodes('&', " &a等級 ： &f" + String.valueOf(p.getLevel())));
-				sender.sendMessage(ChatColor.translateAlternateColorCodes('&', " &a經驗 ： &f" + String.valueOf(Math.floor(p.getExp() * 100) + "%")));
+				sender.sendMessage(ChatColor.translateAlternateColorCodes('&', " &a經驗 ： &f" + String.valueOf((int)Math.floor(p.getExp() * 100) + "%")));
 				sender.sendMessage(ChatColor.translateAlternateColorCodes('&', " &a生命 ： &f" + String.valueOf(p.getHealth()) + " / " + String.valueOf(p.getMaxHealth())));
 				sender.sendMessage(ChatColor.translateAlternateColorCodes('&', " &a魔力 ： &f" + String.valueOf(Mana.getManaBar(p).getMana()) + " / " + String.valueOf(Mana.getManaBar(p).getMaxMana())));
 				sender.sendMessage(ChatColor.translateAlternateColorCodes('&', " &aStr ： &f" + Str));
